Add Camera

+++ app/src/main/java/com/example/mmue_lm3/gameobjects/ItemObject.java
@@ -4,7 +4,7 @@ public abstract class ItemObject extends GameObject {
 
     private static final int PRIORITY = 1;
 
-    protected ItemObject(int x, int y, int width, int height) {
+    protected ItemObject(double x, double y, int width, int height) {
         super(x, y, width, height, PRIORITY);
     }
 }
